general: Clean up signal handler disconnection
Make use of GLib's new g_clear_signal_handler() function to simplify nautilus'
signal handler disconnection and reduce unneccessary code.

Closes: #1522

+++ eel/eel-canvas.h
@@ -355,7 +355,7 @@ struct _EelCanvas {
 	guint idle_id;
 
 	/* Signal handler ID for destruction of the root item */
-	guint root_destroy_id;
+	gulong root_destroy_id;
 
 	/* Internal pixel offsets when zoomed out */
 	int zoom_xofs, zoom_yofs;
